rename assert_matches module
